test(pee): Add jest as devdependency and implement testing suite for mjs and cjs
There are 3 out of the 4 expects that currently fail and are marked with //TODO as I believe they revealed bugs in the workings of the pee
