+++ package.json
@@ -13,7 +13,9 @@
     }
   },
   "scripts": {
-    "test": "node ./test/test.node.js"
+    "test:cjs": "jest --env node --testTimeout 15000 --moduleFileExtensions js --moduleFileExtensions cjs --testMatch '**/*.test.cjs'",
+    "test:mjs": "NODE_OPTIONS=--experimental-vm-modules jest --env jsdom --transform '{}' --testTimeout 15000 --testMatch '**/*.test.js'",
+    "test": "npm run test:cjs && npm run test:mjs"
   },
   "repository": {
     "type": "git",
@@ -40,5 +42,8 @@
   "homepage": "https://github.com/enbyautumn/pee.js#readme",
   "dependencies": {
     "@canvas/image-data": "^1.0.0"
+  },
+  "devDependencies": {
+    "jest": "^27.5.1"
   }
 }
