--- test/index.html
@@ -1,5 +1,0 @@
-<script src="test.js" type="module"></script>
-<!-- Load the test script -->
-<h1>Open the Debugger and use the performance tab to monitor memory usage</h1>
-
-<!-- Usage: Run a web server from the package root folder, then open /test in your browser. -->
