--- test/test.js
@@ -1,3 +1,0 @@
-import { leak } from '../pee.js';
-leak(69);
-// Check in the debugger to see if it is working.
